+++ storey/flow.py
@@ -320,7 +320,8 @@ class Flow:
             result += f"processing_time={event.processing_time}, "
         if getattr(event, "path", None):
             result += f"path={event.path}, "
-        result += f"body={event.body})" if hasattr(event, "body") else ""
+        if getattr(event, "body", None):
+            result += f"body={event.body})"
         return result
 
     def _should_terminate(self):
